// src/app/generated/model/taskDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserDto } from './userDto';
import { TagDto } from './tagDto';


export interface TaskDto { 
    id?: number;
    title?: string;
    description?: string;
    status?: TaskDto.StatusEnum;
    priority?: TaskDto.PriorityEnum;
    dueDate?: string;
    assignee?: UserDto;
    tags?: Array<TagDto>;
    createdAt?: string;
    updatedAt?: string;
}
export namespace TaskDto {
    export const StatusEnum = {
        Todo: 'TODO',
        InProgress: 'IN_PROGRESS',
        InReview: 'IN_REVIEW',
        Done: 'DONE'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
    export const PriorityEnum = {
        Low: 'LOW',
        Medium: 'MEDIUM',
        High: 'HIGH',
        Urgent: 'URGENT'
    } as const;
    export type PriorityEnum = typeof PriorityEnum[keyof typeof PriorityEnum];
}