// src/app/generated/model/taskRequest.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TaskRequest { 
    title: string;
    description?: string;
    status?: TaskRequest.StatusEnum;
    priority?: TaskRequest.PriorityEnum;
    dueDate?: string;
    assigneeId?: number;
    tagIds?: Array<number>;
}
export namespace TaskRequest {
    export const StatusEnum = {
        Todo: 'TODO',
        InProgress: 'IN_PROGRESS',
        InReview: 'IN_REVIEW',
        Done: 'DONE'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
    export const PriorityEnum = {
        Low: 'LOW',
        Medium: 'MEDIUM',
        High: 'HIGH',
        Urgent: 'URGENT'
    } as const;
    export type PriorityEnum = typeof PriorityEnum[keyof typeof PriorityEnum];
}